\name{RDMTkAnalysisR-package}
\alias{RDMTkAnalysisR-package}
\alias{RDMTkAnalysisR}
\docType{package}
\title{
RDMTk Analysis Backend
}
\description{
RDMTk Analysis related tools written in R
}
\details{
\tabular{ll}{
Package: \tab RDMTkAnalysisR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-02-11\cr
License: \tab GPL\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Vinay B Gavirangaswamy
Maintainer: Vinay B Gavirangaswamy <vinay.b.gavirangaswamy@wmich.edu>

}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
